Change/Fix: (--format-element) Use marker's buffer
When formatting an element outside the element's buffer (e.g. when
mapping --format-element across the results of org-ql), the current buffer
is not the element's source buffer, so functions like org-get-tags-at
will not return the correct results, because they don't use the
marker's buffer.  So now we set the current buffer to the marker's
buffer.

As this package matures, this might become unnecessary in the future,
but it's probably best to keep it for correctness.